Use official mongodb driver  (#1745)
* Use official mongodb driver

I used the mgo driver (because it's easier to use) but it doesn't
support the newer version of mongodb (4+) and also doesn't support the
newer connection string format used exclusively by mongodb atlas
(mongo's cloud offering for mongodb, its freee soooooooooo, we are totes
gonna support it).


* Fix little problems with the mongodb implementation

* A couple tests weren't calling teardown
* Fix connection leak with the test database plugin
* Make timeout configurable so that if someone is on a slow machine,
they can wait longer.


* Use int instead of duration for the timeout

+++ pkg/storage/store.go
@@ -14,7 +14,7 @@ type Store interface {
 
 	// Count the number of results that match an optional query.
 	// When the query is omitted, the entire collection is counted.
-	Count(collection string, opts CountOptions) (int, error)
+	Count(collection string, opts CountOptions) (int64, error)
 
 	// EnsureIndex makes sure that the specified index exists as specified.
 	// If it does exist with a different definition, the index is recreated.
